started GaConjOrbit

+++ Utils/Sortable/SortableResult.cs
@@ -12,7 +12,7 @@ namespace Utils.Sortable
             var _sortResults = SortResults.ToList();
             Sorter = _sortResults[0].Sorter;
             Sortable = _sortResults[0].Sortable;
-            float totNess = _sortResults.Sum(r => r.Sortedness);
+            double totNess = _sortResults.Sum(r => r.Sortedness);
             AverageSortedness = totNess / _sortResults.Count;
             StageUse = new int[Sorter.StageCount];
             foreach (var result in _sortResults)
